Add the image-cleanup script!
This is the thing we're actually running!  It calls the Docker API, gets
a list of images, and examines them.  Images which contain at least one
tag are ignored.  Images which contain zero tags are deleted.  We use
Systemd journald logging if available, otherwise we log to stderr.  We
check for Docker exceptions when connecting, and when deleting an image.
A connection exception is fatal; image-deletion exceptions are not.