Start a workflow to run on pushes
This workflow attempts to build a Debian changelog from the history
since the last tag (one 'version' per commit), and then does a `debuild`
without signing.  The (unsigned) results are uploaded as an artifact of
the workflow run.

On the changelog: The Debian package version number is taken from the
changelog.  We need to make at least one changelog entry, as the package
version will be taken from that.  We use version 99, plus the commit
date and tag.  That ensures version numbers continue to increase as we
move forward in time, while also containing the commit ID in the version
number.  If we ever do a proper release of a version above 99, this will
break!

NOTE: `--force-bad-version` is included to allow `debchange` to
continue, even if time does not appear to be moving foward.  This can
happen sometimes, when rebases are performed: You'll have a series of
commits that have the same commit timestamp (the timestamp of the
rebase).