Update workflow to sign our Debian packages
The `package-deb` artifact from our existing workflow contains not just
the Debian package (the `.deb` file), but also all of the files needed
to upload the package into a repository.  But, the files aren't signed,
and most repositories only want uploads that are signed by a trusted
key.

So, this new job does that!  It takes the `package-deb` artifact, uses
`debsign` to sign the appropriate files, and uploads everything to a new
artifact, named `signed-deb`.  This new artifact contains the `.deb`
package files, so you should probably be using this artifact, when it is
available.

The workflow has a few requirements:

* The variable `DEBSIGN_KEYID`, which contains the ID (short or long) of
  the PGP key used for signing.

* The secret `KEY`, which is the armored PGP private key.

* The environment `sign`, containing the secret and variable above.

The job is set to run only on pushes to tags, and to the main branch.